Remove deprecated Bloc events and update unsubscribe logic.
The `ToggleEpisodesFilterStatusEvent` and `RefreshEpisodesByFeedIdEvent` have been removed from the `PodcastBloc`.

When unsubscribing from a podcast, trending podcasts are now fetched again to ensure the remote version of the unsubscribed podcast is displayed if it's in the trending list.

The logic for setting `episodesFilterStatus` and `filterText` when a podcast is tapped has also been removed.