Reset currently playing episode index on app detach.
The `didChangeAppLifecycleState` method now resets the `currentPlayingIndex` in the `PlaylistDetailsCubit` to `null` when the app enters the `AppLifecycleState.detached` state.

This ensures that the playback state is correctly initialized when the app is restarted.

Additionally, the `BackdropFilter` widget used in the error dialog has been replaced with the custom `BackdropFilterWidget`.